+++ packages/rum-core/src/domain/contexts/featureFlagContext.ts
@@ -1,11 +1,5 @@
 import type { RelativeTime, ContextValue, Context, CustomerDataTracker } from '@datadog/browser-core'
-import {
-  noop,
-  isExperimentalFeatureEnabled,
-  SESSION_TIME_OUT_DELAY,
-  ValueHistory,
-  ExperimentalFeature,
-} from '@datadog/browser-core'
+import { SESSION_TIME_OUT_DELAY, ValueHistory } from '@datadog/browser-core'
 import type { LifeCycle } from '../lifeCycle'
 import { LifeCycleEventType } from '../lifeCycle'
 
@@ -32,14 +26,6 @@ export function startFeatureFlagContexts(
   lifeCycle: LifeCycle,
   customerDataTracker: CustomerDataTracker
 ): FeatureFlagContexts {
-  if (!isExperimentalFeatureEnabled(ExperimentalFeature.FEATURE_FLAGS)) {
-    return {
-      findFeatureFlagEvaluations: () => undefined,
-      addFeatureFlagEvaluation: noop,
-      stop: noop,
-    }
-  }
-
   const featureFlagContexts = new ValueHistory<FeatureFlagContext>(FEATURE_FLAG_CONTEXT_TIME_OUT_DELAY)
 
   lifeCycle.subscribe(LifeCycleEventType.BEFORE_VIEW_CREATED, ({ startClocks }) => {